requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
exceptiongroup==1.3.1
filelock==3.20.0
Flask==3.1.3
Flask-PyMongo==3.0.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
Pygments==2.19.2
pymongo==4.16.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.19
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.3
shellingham==1.5.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.2.2+cpu
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
Werkzeug==3.1.6
tensorflow==2.15.0
Pillow==10.4.0
h5py==3.11.0
opencv-python==4.10.0.84